feat: profiles part 1: registration
Up to this point, the strategy for dealing with multiple claim profiles,
is to try unmarshalling each profile's claim structure in turn and then
see "what sticks".

This change introduces an IProfile interface that associates claims with
a profile  name*. An IProfile is registered and the correct one is then
automatically selected by Evidence when unmarshalling CBOR/JSON.

*note: in order to maintain compatibility with P1, the name is a
 string, however, conceptually it should be an EAT profile (i.e. a URL
 or an OID). This should be kept in mind if we want to push profiles
 down into generic EAT code.

Unmarshalling is now a two-phase process:

1. First, the profile claim is identified, extracted, and used to select
   the right claims structure based on registered profiles. If no
   profile claim is present, then P1 is assumed, as it is the only
   profile for which that claim is optional.
2. The claims structure is then unmarshalled as before.

This change means that profiles need not be hard-coded in this code
base, and can be provided by external code using this library by
defining an appropriate IProfile implementation and registering it with
RegisterProfile(). As a bonus, this also produces cleaner, easier to
parse error messages, as only the relevant profile's errors are now
reported.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>