feat: profiles embedding & example
Add an example of extending profiles through embedding an existing
claims implementation.

In order for marshalling to work correctly, we need to be able to treat
the embedded struct as part of its enclosing struct. Add custom
encoding code to support this (adapted from github.com/veraison/corim
which solves a similar problem).

note: this code is not used in the code base itself (outside the
example), but we need to provide it as part of the API to allow
embedding.

Signed-off-by: Sergei Trofimov <sergei.trofimov@arm.com>